+++ frontend/pages/patient/patient-symptoms-daily.tsx
@@ -1,9 +1,8 @@
 import { useSession, getSession } from "next-auth/react";
 import { USER_ROLES } from "@frontend/utils/constants";
-import PatientFormToFill, {
-    requiredDetails,
-    PatientsFormsToFill,
-} from "@frontend/components/forms/patient-form-to-fill";
+import { requiredDetails,  PatientsFormsToFill} from "@frontend/components/forms/types/types";
+
+import PatientFormToFill from "@frontend/components/forms/patient-form-to-fill";
 import { serverURL } from "@frontend/config/index";
 import { NextPageContext } from "next";
